refactor: extract mobile filter sheet effects
